+++ App/Fields/Options/Scripts.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace MC\App\Fields\Options;
+
+use WordPlate\Acf\Fields\Tab;
+use WordPlate\Acf\Fields\Textarea;
+
+/**
+ * Class Scripts
+ *
+ * @package MC\App\Fields\Options
+ */
+class Scripts
+{
+    /**
+     * Defines fields used within Options tab.
+     *
+     * @return array
+     */
+    public function fields()
+    {
+        return apply_filters(
+            'mc/options/scripts',
+            [
+                Tab::make(__('Scripts', 'mc-starter'))
+                    ->placement('left'),
+                Textarea::make(__('Head Scripts', 'crop-nutrition'), 'head-scripts')
+                    ->rows(10),
+                Textarea::make(__('Body Scripts', 'crop-nutrition'), 'body-scripts')
+                    ->rows(10),
+                Textarea::make(__('Footer Scripts', 'crop-nutrition'), 'footer-scripts')
+                    ->rows(10),
+            ]
+        );
+    }
+}